requirements.txt
watchdog>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
Pillow>=10.0.0
customtkinter>=5.2.0
winotify>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
